Partial cleanup of evaluation code
